chore: restructure auth tests
